func_8005AE68 22 -> 21 on one declaration, and two negatives worth keeping
The census as a ranking key paid for itself the first time it ran. `lim`
was s32 and holds only 0x1F, used as `lim - rgb.x` which promotes to int,
so narrowing it is free -- and as s32 the base emitted a VARIABLE shift
(srav) where retail has a constant sra. `u8 lim` removes that pair from the
census, four opcodes to two. On the old (length, differences) ordering this
row read as 22 -> 26 and the scorer would never have surfaced it.

func_80072A48's best stored output reaches an EMPTY census and is still
wrong. Its whole gain is moving `off += 0xC;` inside the body's conditional,
so the cursor would advance only when the test passes. The listing settles
it rather than intuition: that addiu sits in the DELAY SLOT of the loop-back
branch at 80072D08, and the label every early-out jumps to is before that
branch, so a delay slot executes on every path through the body. Retail
advances unconditionally. Its four legitimate edits decompose to exactly
nothing. An empty census says the opcode multiset matches; a cursor that
advances on the wrong path executes the same opcodes.

func_800279BC's own output (14 -> 13) double-applies `(s8)r - 4` to a value
that already carries it. Rejected; `k -= 1;` is worth nothing and four
legitimate spellings of the register merge it stumbled onto are 14, 15, 15
and 32.

Scorer run: 27 functions, 10 with better stored output. Build green at
84747e64.

+++ parked/func_8005AE68.c
@@ -20,7 +20,7 @@ s32 func_8005AE68(s32 arg0, s32 arg1, s32 arg2) {
     Hsv hsv;
     Rgb3 rgb;
     u32 c;
-    s32 lim;
+    u8 lim;
     s32 f;
     s32 h;
     u8 m;
